+++ src/experimental/nodtml/__init__.py
@@ -7,8 +7,15 @@ import os
 
 
 logger = logging.getLogger('experimental.nodtml')
+# Show original dtml text in the logs?
 SHOW = os.environ.get('SHOW_ORIGINAL_DTML')
-VALUE = os.environ.get('DEBUG_DTML_VALUE', u'')
+# Value to use instead of the original dtml text.
+# Note: this must be a string, not unicode.
+# Otherwise you may get exceptions in Products.ResourceRegistries,
+# at least if you have css in dtml files, for example
+# plonetheme/sunburst/skins/sunburst_styles/ploneCustom.css.dtml
+# even though everything there is commented out.
+VALUE = os.environ.get('DEBUG_DTML_VALUE', '')
 
 
 def quotedHTML(self, *args, **kwargs):