+++ CHANGES.rst
@@ -2,8 +2,8 @@ Changelog
 =========
 
 
-1.1.1 (unreleased)
-------------------
+1.2 (unreleased)
+----------------
 
 Breaking changes: